// src/shared/enums/supported-bank-types.ts
export enum SupportedBankTypes {
    AffinBank = 'affin_bank',
    AffinIslamicBank = 'affin_islamic_bank',
    Agrobank = 'agrobank',
    AllianceBank = 'alliance_bank',
    AllianceIslamicBank = 'alliance_islamic_bank',
    AlRajhi = 'al_rajhi',
    Ambank = 'ambank',
    Amislamic = 'amislamic',
    BankIslam = 'bank_islam',
    BankKerjasama = 'bank_kerjasama',
    BankMuamalat = 'bank_muamalat',
    BankOfAmerica = 'bank_of_america',
    BankOfChina = 'bank_of_china',
    BankSimpanan = 'bank_simpanan',
    BankOfTokyoMitsubishiUfj = 'bank_of_tokyo_mitsubishi_ufj',
    CimbBank = 'cimb_bank',
    CimbIslamicBank = 'cimb_islamic_bank',
    BnpParibas = 'bnp_paribas',
    Citibank = 'citibank',
    DeutscheBank = 'deutsche_bank',
    HongLeongBank = 'hong_leong_bank',
    HongLeongIslamicBank = 'hong_leong_islamic_bank',
    HsbcBank = 'hsbc_bank',
    HsbcAmanah = 'hsbc_amanah',
    IndustrialCommercialBankOfChina = 'industrial_commercial_bank_of_china',
    JpMorganChase = 'jp_morgan_chase',
    KuwaitFinanceHouse = 'kuwait_finance_house',
    Maybank = 'maybank',
    MaybankIslamic = 'maybank_islamic',
    MizuhoBank = 'izuho_bank',
    OcbcBank = 'ocbc_bank',
    OcbcAlAminBank = 'ocbc_al_amin_bank',
    PublicBank = 'public_bank',
    PublicIslamicBank = 'public_islamic_bank',
    RhbBank = 'rhb_bank',
    RhbIslamicBank = 'rhb_islamic_bank',
    StandardCharteredBank = 'standard_chartered_bank',
    StandardCharteredSaadiq = 'standard_chartered_saadiq',
    SumitomoMitsuiBankingCorporation = 'sumitomo_mitsui_banking_corporation',
    UnitedOverseasBank = 'united_overseas_bank',
}